chore(meshupgrade): refactor function places
