+++ plugins/lime-plugin-mesh-wide-upgrade/src/mesWideUpgradeQueries.tsx
@@ -3,6 +3,7 @@ import { useMutation, useQuery } from "@tanstack/react-query";
 import {
     getMeshUpgradeNodeStatus,
     getMeshWideUpgradeInfo,
+    remoteStartSafeUpgrade,
     setBecomeMainNode,
     setStartFirmwareUpgradeTransaction,
 } from "plugins/lime-plugin-mesh-wide-upgrade/src/meshWideUpgradeApi";
@@ -10,6 +11,9 @@ import {
     MeshWideUpgradeInfo,
     NodeMeshUpgradeInfo,
 } from "plugins/lime-plugin-mesh-wide-upgrade/src/meshWideUpgradeTypes";
+import { useMeshWideNodes } from "plugins/lime-plugin-mesh-wide/src/mesWideQueries";
+
+import { useMeshWideSyncCall } from "utils/meshWideSyncCall";
 
 export const meshUpgradeNodeStatusKey = [
     "lime-mesh-upgrade",
@@ -53,3 +57,15 @@ export function useStartFirmwareUpgradeTransaction(params) {
         ...params,
     });
 }
+
+// Parallel queries/mutations
+
+export const useStartSafeUpgrade = () => {
+    // State to store the errors
+    const { data: nodes } = useMeshWideNodes({});
+    const ips = Object.values(nodes || {}).map(
+        // @ts-ignore
+        (node) => node?.ipv4 ?? ""
+    );
+    return useMeshWideSyncCall(remoteStartSafeUpgrade, ips);
+};
